ariadne_server/requirements.txt
ariadne
httpx
requests
pytest-asyncio
uvicorn
aioredis
grpcio
grpcio-tools
googleapis-common-protos
argon2-cffi
gino
grpclib
aiostream
pymacaroons
